pyproject.toml
[project]
name = "analytics-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.1",
    "apache-airflow-providers-google>=18.1.0",
    "dbt-bigquery>=1.10.2",
    "faker>=37.12.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=3.1.1",
    "ipykernel>=7.1.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
]
